Show only active experiment content. Closes #2

+++ mod/reservations/experiments/index.php
@@ -34,18 +34,26 @@ if (!$lab){
     <?php
   $experiments = $lab->experiments();
   foreach ($experiments as $experiment) {
-    echo "<tr>";
-    echo "<td>" . $experiment->name . "</td>";
-    echo "<td><a href='../contents/index.php?experiment_id=" . $experiment->id . "'>Ver Documentos</a></td>";
     if (has_capability("mod/reservations:update_experiment", $context)) {
+      echo "<tr>";
+      echo "<td>" . $experiment->name . "</td>";
+      echo "<td><a href='../contents/index.php?experiment_id=" . $experiment->id . "'>Ver Documentos</a></td>";
       echo "<td>". $experiment->activation_link($OUTPUT->pix_url('i/show'), $OUTPUT->pix_url('i/hide')) . "</td>";
       echo "<td><a href='../show.php?id=0&exp=" . $experiment->id . "'><img title='Vista Previa' alt='Vista Previa' src='" . $OUTPUT->pix_url('f/edit') . "'/></a></td>";
       echo "<td><a href='edit.php?experiment_id=" . $experiment->id . "'><img title='Editar' alt='Editar' src='" . $OUTPUT->pix_url('a/setting') . "'/></a></td>";
+      if (has_capability("mod/reservations:delete_experiment", $context)) {
+        echo "<td><a href='delete.php?experiment_id=" . $experiment->id . "' onclick='return confirm(\"Una vez borrado el experimento no se puede recuperar. Seguro que desea eliminar el experimento?\")'><img title='Eliminar' alt='Eliminar' src='" . $OUTPUT->pix_url('i/cross_red_big') . "'/></a></td>";
+      }
+      echo "</tr>";
     }
-    if (has_capability("mod/reservations:delete_experiment", $context)) {
-      echo "<td><a href='delete.php?experiment_id=" . $experiment->id . "' onclick='return confirm(\"Una vez borrado el experimento no se puede recuperar. Seguro que desea eliminar el experimento?\")'><img title='Eliminar' alt='Eliminar' src='" . $OUTPUT->pix_url('i/cross_red_big') . "'/></a></td>";
+    else {
+      if ($experiment->is_active) {
+        echo "<tr>";
+        echo "<td>" . $experiment->name . "</td>";
+        echo "<td><a href='../contents/index.php?experiment_id=" . $experiment->id . "'>Ver Documentos</a></td>";
+        echo "</tr>";
+      }
     }
-    echo "</tr>";
   }
     ?>
   </table>
